+++ src/player/Node.cpp
@@ -336,7 +336,7 @@ void Node::releaseEventCapture(int cursorID) {
 void Node::setEventHandler(Event::Type Type, int Sources, PyObject * pFunc)
 {
     for (int i=0; i<4; ++i) {
-        int source = pow(2.,i);
+        int source = int(pow(2.,i));
         if (source & Sources) {
             EventHandlerID ID(Type, (Event::Source)source);
             EventHandlerMap::iterator it = m_EventHandlerMap.find(ID);
